Added ability to delete nodes and properties

+++ Munyn/ViewModels/Nodes/NodeBaseViewModel.cs
@@ -99,6 +99,7 @@ namespace Munyn.ViewModels
         [RelayCommand]
         private void DeleteNode()
         {
+            if (_mainVM != null)
             _mainVM.DeleteNode(this);
         }
